[REF] l10n_br_fiscal: document_serie_id ->compute
